+++ src/billing/errors/index.ts
@@ -1,1 +1,2 @@
 export { default as UnableToContinueAsGuestError } from './unable-to-continue-as-guest-error';
+export { default as UpdateCustomerError } from './update-customer-error';