+++ src/customer/customer-request-sender.ts
@@ -1,7 +1,8 @@
 import { RequestSender, Response } from '@bigcommerce/request-sender';
 
-import { RequestOptions } from '../common/http-request';
+import { ContentType, RequestOptions } from '../common/http-request';
 
+import { CurrentCustomer } from './customer';
 import CustomerCredentials from './customer-credentials';
 import { InternalCustomerResponseBody } from './internal-customer-responses';
 
@@ -9,6 +10,13 @@ export default class CustomerRequestSender {
     constructor(
         private _requestSender: RequestSender
     ) {}
+
+    updateCustomer(customer: CurrentCustomer, { timeout }: RequestOptions = {}): Promise<Response<CurrentCustomer>> {
+        const url = '/api/storefront/customer';
+        const headers = { Accept: ContentType.JsonV1 };
+
+        return this._requestSender.post(url, { body: customer, headers, timeout });
+    }
 
     signInCustomer(credentials: CustomerCredentials, { timeout }: RequestOptions = {}): Promise<Response<InternalCustomerResponseBody>> {
         const url = '/internalapi/v1/checkout/customer';