+++ src/customer/guest-credentials.ts
@@ -1,4 +1,5 @@
 export default interface GuestCredentials {
     id?: string;
     email: string;
+    marketingEmailConsent?: boolean;
 }